added authorization for modifier functions

+++ shoppingList/delete.js
@@ -1,9 +1,10 @@
 const AWS = require("aws-sdk");
+const decodeVerify = require("../auth/decodeVerify");
 
 const documentClient = new AWS.DynamoDB.DocumentClient();
 
 module.exports.main = async event => {
-  const { id, timestamp } = JSON.parse(event.body);
+  const { username, idToken, id, timestamp } = JSON.parse(event.body);
   const params = {
     TableName: process.env.SHOPPING_LIST_TABLE,
     Key: {
@@ -12,16 +13,35 @@ module.exports.main = async event => {
     }
   };
 
-  return new Promise(resolve => {
-    documentClient.delete(params, err => {
-      if (err) {
-        resolve({
-          statusCode: 501,
-          message: err.message || "Could not delete an item."
+  return await new Promise(resolve => {
+    decodeVerify(idToken)
+      .then(({ email }) => {
+        if (email !== username) {
+          resolve({
+            statusCode: 401,
+            body: JSON.stringify({ message: "Unauthorized operation." })
+          });
+          return;
+        }
+
+        documentClient.delete(params, err => {
+          if (err) {
+            resolve({
+              statusCode: 501,
+              body: JSON.stringify({
+                message: err.message || "Could not delete an item."
+              })
+            });
+          } else {
+            resolve({ statusCode: 200, body: JSON.stringify({ id }) });
+          }
         });
-      } else {
-        resolve({ statusCode: 200, body: JSON.stringify({ id }) });
-      }
-    });
+      })
+      .catch(e => {
+        resolve({
+          statusCode: 401,
+          body: JSON.stringify({ message: "Unauthorized operation." })
+        });
+      });
   });
 };
